refinement

+++ plotter.py
@@ -24,28 +24,21 @@ label_dict = {
 
 
 def plot_pie(d, index):
+    fname = os.path.join(info_dir, label_dict[index] + '.png')
+    if os.path.exists(fname):
+        return
     size = d.shape[0]
     v_c_dict = collections.Counter(d)
-    v_p_dict = {v: c / float(size) for v, c in v_c_dict.items()}
+    v_p_dict = {v: c / float(size) * 100 for v, c in v_c_dict.items()}
     # labels = [gender_dict[v] for v in v_p_dict.keys()]
-    labels = v_p_dict.keys()
+    # labels = v_p_dict.keys()
     values = v_p_dict.values()
-    plt.pie(
-        values,
-        labels=labels,
-        autopct='%.1f%%',
-        shadow=False,
-        startangle=90)
-    plt.axis('equal')
-    title = label_dict[index] + ' total={}'.format(size)
-    # plt.title(title, loc='center')
+    patches, texts = plt.pie(values, shadow=False, startangle=90)
+    # plt.axis('equal')
+    title = label_dict[index].upper() + ' total={}'.format(size)
+    # legend
+    legend_labels = ['{:.2f}%'.format(v) for v in values]
+    plt.legend(patches, legend_labels, loc='best')
     plt.title(title, bbox={'facecolor': '0.8', 'pad': 5}, loc='right')
-    fname = os.path.join(info_dir, label_dict[index] + '.png')
     plt.savefig(fname)
     plt.clf()
-
-
-if __name__ == '__main__':
-    d = dataset.DataReader()
-    for i in [0, 1, 4, 5]:
-        plot_pie(d, i)
